Issue #10981:add delete confirmation prompts
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@8677 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/customer.cpp
@@ -938,6 +938,13 @@ void customer::sViewShipto()
 
 void customer::sDeleteShipto()
 {
+  QString question = tr("Are you sure that you want to delete this Ship To?");
+  if (QMessageBox::question(this, tr("Delete Ship To?"),
+                              question,
+                              QMessageBox::Yes,
+                              QMessageBox::No | QMessageBox::Default) == QMessageBox::No)
+    return;
+
   q.prepare("SELECT deleteShipTo(:shipto_id) AS result;");
   q.bindValue(":shipto_id", _shipto->id());
   q.exec();
